# docs/api.yml
openapi: 3.1.0
info:
  title: Food Bank Finder & Management API
  version: 1.0.0
  description: jade-jasmine team api
paths:
...
  /foodbank:
    get:
      tags:
        - completed
        - food bank
      summary: Get food bank list
      security: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - name
                    - street
                    - city
                    - country
                    - charity_id
                    - admin
                  properties:
                    id:
                      type: integer
                    fax:
                      type: integer
                    city:
                      type: string
                    name:
                      type: string
                    admin:
                      $ref: "#/components/schemas/User"
                    phone:
                      type: integer
                    street:
                      type: string
                      examples:
                        - 555 Food Bank Way
                        - 55 Pantry St
                    country:
                      type: string
                    unit_no:
                      type: string
                      description: Can be a number or a letter
                    website:
                      type: string
                    latitude:
                      type: number
                      format: float
                    province:
                      type: string
                    longitude:
                      type: number
                      format: float
                    charity_id:
                      type: string
                    postal_code:
                      type: string
                    email:
                      type: string
                    description:
                      type: string
                  description: The food bank or food pantry entity and all its details.
                title: GetFoodBankListOk
          description: The request was successful, and the server has returned the
            requested resource in the response body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters:
        - $ref: "#/components/parameters/city"
        - $ref: "#/components/parameters/fb_name"
        - in: query
          name: fb_id
          style: form
          schema:
            type: integer
          explode: true
          required: false
          description: the id of the food bank
          allowReserved: false
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      description: >-
        Fetches a comprehensive list of all food banks in our db. Each food bank
        object includes essential details but doesn't include hours or staff.

        Also doesn't include the admin id as that is protected information.


        If the request is made by an authenticated user who also happens to be
        the admin of the same food bank, then it will return the admin's
        username and id as well as their email address.
      operationId: getFoodBankList
    post:
      tags:
        - completed
      summary: Create food bank
      security:
        - BearerJwt: []
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodBank"
          description: The request was successful, and a new resource was created.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
          description: Authentication is required to access the requested resource. The
            client must include the appropriate credentials.
        "403":
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
          description: The server understood the request, but refuses to authorize it.
            Ensure the client has appropriate permissions.
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      parameters: []
      description: |-
        must be an authenticated user to create a food bank
        (initially assume a 1-1 relationship between food bank and admin)
      operationId: createFoodBank
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - city
                - province
                - country
                - postal_code
                - phone
                - street
                - timezone
                - description
                - published
                - email
              properties:
                fax:
                  type: integer
                  description: |
                    the fax number
                  examples:
                    - 5555555555
                city:
                  type: string
                  description: the city name
                  examples:
                    - Toronto
                    - Los Angeles
                name:
                  type: string
                  description: the name of the charity
                  examples:
                    - Toronto Food Bank
                phone:
                  type: integer
                  description: the phone number
                  examples:
                    - 5555555555
                street:
                  type: string
                  description: the street name and number
                  examples:
                    - 222 Tomboy St
                    - 33 Old Time Ave.
                country:
                  type: string
                unit_no:
                  type: string
                  description: the unit no in a building
                  examples:
                    - 11a
                website:
                  type: string
                  description: the url of the food bank website
                  examples:
                    - https://website.ca
                province:
                  type: string
                  description: the state/provine or area
                  examples:
                    - Ontario
                    - California
                charity_no:
                  type: string
                  description: the charity number
                  examples:
                    - "19282"
                postal_code:
                  type: string
                timezone:
                  $ref: "#/components/schemas/timezone"
                  description: the IANA timezone value like America/Toronto
                description:
                  type: string
                published:
                  type: string
                email:
                  type: string
        required: true
        description: The food bank's static data. The authorized user will become the
          primary admin for this food bank.
...
components:
  schemas:
...
    User:
      type: object
      required:
        - username
        - email
        - id
      properties:
        id:
          type: integer
          minimum: 1
          examples:
            - 1
            - 2
            - 3
          description: the id of hte user
        email:
          type: string
          examples:
            - admin@gmail.com
          description: the email of the user
        username:
          type: string
          examples:
            - panther
            - cobra
          description: the username of the user
      description: A user is either a food bank admin or a food bank employee (note
        they are all considered staff)
...
    FoodBank:
      type: object
      required:
        - id
        - street
        - city
        - country
        - hours
        - admin
        - timzone
        - published
        - fbname
      properties:
        id:
          type: integer
          description: auto generated id
          minimum: 1
          examples:
            - 1
            - 2
            - 3
        fax:
          type: integer
          description: the fax number for this food bank
        city:
          type: string
        admin:
          $ref: "#/components/schemas/User"
        hours:
          type: string
        phone:
          type: integer
        staff:
          type: array
          items:
            type: integer
            minimum: 1
            description: the id of the employee
        street:
          type: string
          examples:
            - 555 Food Bank Way
            - 55 Pantry St
        country:
          type: string
        timzone:
          type: string
          examples:
            - America/Toronto
          description: this value is the IANA timezone value
        unit_no:
          type: string
          examples:
            - 121a
          description: Can be a number or a letter
        website:
          type: string
          examples:
            - info@website.com
          description: the food bank's website address
        latitude:
          type: number
          format: float
        province:
          type: string
        longitude:
          type: number
          format: float
        postal_code:
          type: string
        charity_registration_no:
          type: string
        published:
          type: boolean
          description: defaults to true
          examples:
            - false
        fbname:
          type: string
          examples:
            - Fancy Food Bank
          description: the name of the food bank
        email:
          type: string
          description: the email address of the food bank
          examples:
            - hana@email.com
        description:
          type: string
      description: The food bank or food pantry entity and all its details.
    timezone:
      type: string
      description: the IANA timezone string
      examples:
        - America/Toronto
  responses:
    Conflict:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                examples:
                  - Lorem ipsum
      description: The request could not be completed due to a conflict with the
        current state of the resource. Resolve the conflict and try again.
...
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                examples:
                  - Lorem ipsum
                description: the error message
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
    InternalServerErrror:
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                examples:
                  - Lorem ipsum
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
  parameters:
...
    city:
      in: query
      name: city
      style: form
      schema:
        type: string
        examples:
          - toronto
          - metropolis
        description: the name of the city
      explode: true
      description: the name of a city
      allowReserved: false
      allowEmptyValue: false
...
    limit:
      name: limit
      in: query
      description: |-
        the number of items to return per page of results
        Default is 50 (min is 1)
        Field is automatically sanitized
      schema:
        type: integer
      explode: true
      style: form
      allowReserved: false
      allowEmptyValue: false
...
    offset:
      name: offset
      in: query
      description: >-
        the number of rows to offset to (eg. if offset is 20, the first row
        retrieved will be row 20)

        Default is 0. (field is automatically sanitized in case of invalid
        values)
      schema:
        type: integer
      explode: true
      style: form
      allowReserved: false
      allowEmptyValue: false
    fb_name:
      in: query
      name: fb_name
      style: form
      schema:
        type: string
        examples:
          - Charity Food
          - Community Food Bank
        description: the name of the food bank
      explode: true
      description: the name of the food bank
      allowReserved: false
      allowEmptyValue: false
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
      bearerFormat: Json Web Token (JWT)
...